do_u_even_ransom/requirements.txt
cryptography
pycryptodome
requests
